// backend/src/common/enums/enums.ts
export enum EstadoPostulacion {
  postulado = 'postulado',
  aceptado = 'aceptado',
  rechazado = 'rechazado',
  en_curso = 'en_curso',
  finalizado = 'finalizado',
}

export enum EstadoTesis {
  propuesta = 'propuesta',
  desarrollo = 'desarrollo',
  sustentacion = 'sustentacion',
  culminado = 'culminado',
}

export enum RolNombre {
  admin = 'admin',
  coordinador = 'coordinador',
  asesor = 'asesor',
  estudiante = 'estudiante',
  empresa = 'empresa',
}